New Functions and organization
Added two functions that encapsulate functionality that was achieved by
this fucked up if/then tree. One function reformats input equations, and
the other checks to see if they're evaluable. Also, considering banging
together a little expression library of my own to use for this, since
NCalc is kinda clunky.