New parser added
I wrote a bunch of these little Token structs (and discovered that structs can't inherit from classes? I wanted a single ToString() method for them but maybe that can come later) and a first pass of the tokenizer and parser. The tokenzier doesn't really... work, yet, I think. It spits out too many tokens.

The next task: Testing! I'm gonna add a testing project to exercise this and make the bugs go away. After that, I'll integrate it into the actual app and replace NCalc and clean up a bunch of my old, nasty code. The last commit on that was like 1100 days ago. Jesus.