Corrected test cases and fixed a bug
The test cases with parenthesis were incorrectly written. I found a website that produces (presumably) correct output, so I'm testing against that.

I also corrected a bug that assumed that the item at the top of the operators stack was always an operator. Sometimes, it is a parenthesis.

The condition is very long so I'm gonna investigate cleaning it up. I dunno if I can do it cleanly.