Finished fixing RegexTokenizer, and (hopefully) got settings settled
RegexTokenizer now correctly tokenizers expressions. It reads any negative signs (no matter the spacing) that are preceded by a number as an Operator -. If it is preceded by a non-number, then it is read as part of the number, to make it explicitly negative. This is, I think, more sensible, since "1 -3" doesn't make much sense as "One, Negative Three".

RegexTokenizer also emits EndOfExpressionTokens again. This was a previous behavior that I removed, since I didn't see the usefulness. But it's marking works well, and prevents CurrentToken from being null except before reading any tokens. (I might add a BeginningOfExpressionToken type later and have the Tokenizer start there).

So, I added EndOfExpression tokens to the end of all the test items, and adjusted some test data to make it test the actual features.

Finally, there's a few assorted changes to some files to get the solution loading and building okay. Some of the problems were just Rider being overly picky, or set in a weird way. It all seems to build alright.

The next tasks: Implement expression evaluation for the Expression class, and variable tokenizing in RegexTokenizer. Since I have that working, I might drop IterativeTokenizer, since that was intended as a replacement to handle the minus/negative-sign mixup.